+++ alembic/env.py
@@ -4,7 +4,7 @@ from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 
 from alembic import context
-from bookscraper_backend.database.db_models import Base
+from src.database.db_models import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
